// apps/web/src/proxy/models/authenticationMethod.ts
// AUTO-GENERATED — DO NOT EDIT.
// Regenerate with: npm run generate:proxy
// Source: StarterKit.WebApi | v1 1.0.0

export type AuthenticationMethod = (typeof AuthenticationMethod)[keyof typeof AuthenticationMethod];

export const AuthenticationMethod = {
  Credentials: 'Credentials',
  Microsoft365: 'Microsoft365',
  Google: 'Google',
  PhoneOtp: 'PhoneOtp',
  CustomAuthentication: 'CustomAuthentication',
} as const;