// apps/web/src/proxy/models/setupStatus.ts
// AUTO-GENERATED — DO NOT EDIT.
// Regenerate with: npm run generate:proxy
// Source: StarterKit.WebApi | v1 1.0.0

export type SetupStatus = (typeof SetupStatus)[keyof typeof SetupStatus];

export const SetupStatus = {
  None: 'None',
  PendingSetup: 'PendingSetup',
  SetupExpired: 'SetupExpired',
} as const;